✨ Support for PSResourceGet, RequiredModules, and PSDepend manifest files
The `-Path` parameter can now be directed to a manifest .psd1 or .json file of these alternate specification formats and ModuleFast will resolve and install the packages.

ModuleFast has some limited autodetection for the format used, but in case it guesses wrong, there is also a new `-SpecFileType` parameter to explicitly define which format the file is in. Check the debug messages to see which format it detected and why.